[#2697 #2706] Move SearchMarkerUtils.js into Models/ ahead of refactoring.
