+++ lib/Models/SearchMarkerUtils.js
@@ -6,7 +6,7 @@ import Cartesian3 from 'terriajs-cesium/Source/Core/Cartesian3';
 import VerticalOrigin from 'terriajs-cesium/Source/Scene/VerticalOrigin';
 import sampleTerrain from 'terriajs-cesium/Source/Core/sampleTerrain';
 
-import markerIcon from '../../../wwwroot/images/map-pin.png';
+import markerIcon from '../../wwwroot/images/map-pin.png';
 
 const SEARCH_MARKER_DATA_SOURCE_NAME = 'TerriaJS Search Marker Points';
